[#11900] Move KerberosCatalogUtils to iceberg-common utils package
Relocate Kerberos catalog helpers alongside other catalog utilities while keeping KerberosConfig in authentication.kerberos.
